Tweak the extraction process.
Update which SQL files need to be introspected for generating the EdgeQL for
the extension. Clean up the template, since the `<>` is now properly
implemented and we don't need to fake it.

Remove debug print from the tests.

Include the generated .rst in the repo (so that it can be potentially
used by the docs).